chore: fix price impact issue
